refactor: introduce data_types module and new module structure
