+++ Cargo.toml
@@ -23,10 +23,9 @@ generate_new_bindings_to_graphblas_implementation = []
 bindgen = "0.58"
 cmake = "0.1"
 
-[build]
-
 [dev-dependencies]
 criterion = "0.3"
+rayon = {version = "1.5"}
 
 [[bench]]
 name = "insert_scalar_into_matrix"
