+++ src/lib.rs
@@ -2,8 +2,5 @@ mod bindings_to_graphblas_implementation;
 pub mod context;
 pub mod error;
 pub mod operators;
-pub mod sparse_matrix;
-pub(crate) mod sparse_scalar;
-pub mod sparse_vector;
 pub mod util;
-pub mod value_type;
+pub mod value_types;
