+++ src/operators/element_wise_multiplication/element_wise_matrix_multiplication.rs
@@ -6,8 +6,8 @@ use crate::operators::{
     binary_operator::BinaryOperator, mask::MatrixMask, monoid::Monoid, options::OperatorOptions,
     semiring::Semiring,
 };
-use crate::sparse_matrix::SparseMatrix;
-use crate::value_type::{AsBoolean, ValueType};
+use crate::value_types::sparse_matrix::SparseMatrix;
+use crate::value_types::value_type::{AsBoolean, ValueType};
 
 use crate::bindings_to_graphblas_implementation::{
     GrB_BinaryOp, GrB_Descriptor, GrB_Matrix_eWiseMult_BinaryOp, GrB_Matrix_eWiseMult_Monoid,
@@ -279,7 +279,7 @@ mod tests {
 
     use crate::context::{Context, Mode};
     use crate::operators::binary_operator::{First, Plus, Times};
-    use crate::sparse_matrix::{
+    use crate::value_types::sparse_matrix::{
         FromMatrixElementList, GetMatrixElementList, GetMatrixElementValue, MatrixElementList, Size,
     };
