+++ src/operators/extract/extract_sub_matrix.rs
@@ -5,11 +5,11 @@ use crate::error::SparseLinearAlgebraError;
 use crate::operators::{
     binary_operator::BinaryOperator, mask::MatrixMask, options::OperatorOptions,
 };
-use crate::sparse_matrix::SparseMatrix;
+use crate::value_types::sparse_matrix::SparseMatrix;
 use crate::util::{
     ElementIndex, ElementIndexSelector, ElementIndexSelectorGraphblasType, IndexConversion,
 };
-use crate::value_type::{AsBoolean, ValueType};
+use crate::value_types::value_type::{AsBoolean, ValueType};
 
 use crate::bindings_to_graphblas_implementation::{
     GrB_BinaryOp, GrB_Descriptor, GrB_Matrix_extract,
@@ -276,7 +276,7 @@ mod tests {
 
     use crate::context::{Context, Mode};
     use crate::operators::binary_operator::First;
-    use crate::sparse_matrix::{FromMatrixElementList, GetMatrixElementValue, MatrixElementList};
+    use crate::value_types::sparse_matrix::{FromMatrixElementList, GetMatrixElementValue, MatrixElementList};
 
     #[test]
     fn test_matrix_extraction() {
