+++ src/operators/insert/insert_vector_into_vector.rs
@@ -7,9 +7,9 @@ use crate::operators::{
     binary_operator::BinaryOperator, mask::VectorMask, options::OperatorOptions,
 };
 
-use crate::sparse_vector::SparseVector;
+use crate::value_types::sparse_vector::SparseVector;
 use crate::util::{ElementIndexSelector, ElementIndexSelectorGraphblasType, IndexConversion};
-use crate::value_type::{AsBoolean, ValueType};
+use crate::value_types::value_type::{AsBoolean, ValueType};
 
 use crate::bindings_to_graphblas_implementation::{
     GrB_BinaryOp, GrB_Descriptor, GrB_Vector_assign,
@@ -209,7 +209,7 @@ mod tests {
     use crate::context::{Context, Mode};
     use crate::operators::binary_operator::First;
 
-    use crate::sparse_vector::{FromVectorElementList, GetVectorElementValue, VectorElementList};
+    use crate::value_types::sparse_vector::{FromVectorElementList, GetVectorElementValue, VectorElementList};
     use crate::util::ElementIndex;
 
     #[test]
