+++ src/operators/select/diagonal_index.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use crate::context::Context;
 use crate::error::SparseLinearAlgebraError;
 
-use crate::sparse_scalar::{SetScalarValue, SparseScalar};
+use crate::value_types::sparse_scalar::{SetScalarValue, SparseScalar};
 
 type GraphblasDiagonalDiagonalIndex = i64;
