+++ src/operators/subinsert/insert_matrix_into_sub_matrix.rs
@@ -3,11 +3,13 @@ use std::ptr;
 use std::marker::PhantomData;
 
 use crate::error::SparseLinearAlgebraError;
-use crate::operators::{binary_operator::BinaryOperator, mask::MatrixMask, options::OperatorOptions};
-use crate::sparse_matrix::SparseMatrix;
+use crate::operators::{
+    binary_operator::BinaryOperator, mask::MatrixMask, options::OperatorOptions,
+};
+use crate::value_types::sparse_matrix::SparseMatrix;
 
 use crate::util::{ElementIndexSelector, ElementIndexSelectorGraphblasType, IndexConversion};
-use crate::value_type::{AsBoolean, ValueType};
+use crate::value_types::value_type::{AsBoolean, ValueType};
 
 use crate::bindings_to_graphblas_implementation::{
     GrB_BinaryOp, GrB_Descriptor, GxB_Matrix_subassign,
@@ -312,7 +314,7 @@ mod tests {
     use crate::context::{Context, Mode};
     use crate::operators::binary_operator::First;
 
-    use crate::sparse_matrix::{
+    use crate::value_types::sparse_matrix::{
         FromMatrixElementList, GetMatrixElementValue, MatrixElementList, Size,
     };
     use crate::util::ElementIndex;
