+++ src/value_types/sparse_matrix/element.rs
@@ -1,7 +1,7 @@
 use super::coordinate::Coordinate;
 use crate::error::{GraphBlasError, GraphBlasErrorType, LogicError, LogicErrorType};
 use crate::util::ElementIndex;
-use crate::value_type::ValueType;
+use crate::value_types::value_type::ValueType;
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct MatrixElement<T: ValueType> {
