+++ src/value_types/sparse_vector/sparse_vector.rs
@@ -33,13 +33,41 @@ use crate::bindings_to_graphblas_implementation::{
 use crate::context::Context;
 use crate::operators::binary_operator::BinaryOperator;
 use crate::util::{ElementIndex, IndexConversion};
-use crate::value_type::{BuiltInValueType, CustomValueType, RegisteredCustomValueType, ValueType};
+use crate::value_types::value_type::{BuiltInValueType, CustomValueType, RegisteredCustomValueType, ValueType};
 
 pub struct SparseVector<T: ValueType> {
     context: Arc<Context>,
     vector: GrB_Vector,
     value_type: PhantomData<T>,
 }
+
+// Send and Sync implementaioms should be ok, since mutable access to GrB_Vector 
+// must occur through a mut SparseVector. Method providing a copy or reference to 
+// GrB_Vector will result in undefined behaviour though. Code review must consider this.
+// https://doc.rust-lang.org/nomicon/send-and-sync.html
+// unsafe impl Send for SparseVector<bool> {}
+// unsafe impl Send for SparseVector<u8> {}
+// unsafe impl Send for SparseVector<u16> {}
+// unsafe impl Send for SparseVector<u32> {}
+// unsafe impl Send for SparseVector<u64> {}
+// unsafe impl Send for SparseVector<i8> {}
+// unsafe impl Send for SparseVector<i16> {}
+// unsafe impl Send for SparseVector<i32> {}
+// unsafe impl Send for SparseVector<i64> {}
+// unsafe impl Send for SparseVector<f32> {}
+// unsafe impl Send for SparseVector<f64> {}
+
+// unsafe impl Sync for SparseVector<bool> {}
+// unsafe impl Sync for SparseVector<u8> {}
+// unsafe impl Sync for SparseVector<u16> {}
+// unsafe impl Sync for SparseVector<u32> {}
+// unsafe impl Sync for SparseVector<u64> {}
+// unsafe impl Sync for SparseVector<i8> {}
+// unsafe impl Sync for SparseVector<i16> {}
+// unsafe impl Sync for SparseVector<i32> {}
+// unsafe impl Sync for SparseVector<i64> {}
+// unsafe impl Sync for SparseVector<f32> {}
+// unsafe impl Sync for SparseVector<f64> {}
 
 impl<T: ValueType + BuiltInValueType<T>> SparseVector<T> {
     pub fn new(
